moved filterdropdown componet to seperate file
